Add estate places

+++ app/src/main/java/com/openclassrooms/realestatemanager/data/dao/EstatePlaceDao.kt
@@ -7,7 +7,7 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface EstatePlaceDao {
     @Query("SELECT * FROM estate_place WHERE estate_id = :estateId ORDER BY id")
-    fun getPlacesForAEstate(estateId: Int?): Flow<List<EstatePlace>>
+    fun getPlacesForAEstate(estateId: Long?): Flow<List<EstatePlace>>
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
     suspend fun insert(estatePlace: EstatePlace)
